refactor verify functions
